backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-multipart>=0.0.9
PyMuPDF>=1.24
python-docx>=1.1
